% Auto-generated: do not edit by hand
\name{tree}

\alias{tree}

\title{Tree component}

\description{

}

\usage{
tree(id=NULL, className=NULL, collapse_color=NULL, data=NULL,
height=NULL, open_by_default=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Class name of the tree.}

\item{collapse_color}{Character. A list of dictionaries that defines the tree structure.}

\item{data}{Unnamed list. A list of dictionaries that defines the tree structure.}

\item{height}{Numeric. A list of dictionaries that defines the tree structure.}

\item{open_by_default}{Logical. A list of dictionaries that defines the tree structure.}

\item{width}{Numeric. A list of dictionaries that defines the tree structure.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

