% Auto-generated: do not edit by hand
\name{tree}

\alias{tree}

\title{Tree component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
tree(id=NULL, data=NULL, leaf_icon_color=NULL,
open_by_default=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Unnamed list. A list of dictionaries that defines the tree structure.}

\item{leaf_icon_color}{Character. A list of dictionaries that defines the tree structure.}

\item{open_by_default}{Logical. A list of dictionaries that defines the tree structure.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

